Перевел все приложение на REST контроллеры, немного разобрался с Постманом, все работает!
