+++ src/main/java/ru/itmentor/spring/boot_security/demo/configs/SuccessUserHandler.java
@@ -18,7 +18,7 @@ public class SuccessUserHandler implements AuthenticationSuccessHandler {
         if (roles.contains("ROLE_ADMIN")) {
             response.sendRedirect("/admin/user");
         } else if (roles.contains("ROLE_USER")) {
-            response.sendRedirect("/user");
+            response.sendRedirect("/api/user");
         } else {
             response.sendRedirect("/");
         }
