--- src/main/java/ru/itmentor/spring/boot_security/demo/controllers/RoleRestController.java
@@ -1,44 +1,0 @@
-package ru.itmentor.spring.boot_security.demo.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import ru.itmentor.spring.boot_security.demo.models.Role;
-import ru.itmentor.spring.boot_security.demo.service.RoleService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/roles")
-public class RoleRestController {
-
-    private final RoleService roleService;
-
-    @Autowired
-    public RoleRestController(RoleService roleService) {
-        this.roleService = roleService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<Role>> getAllRoles() {
-        return ResponseEntity.ok(roleService.getAllRoles());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Role> getRoleById(@PathVariable Long id) {
-        Role role = roleService.getRoleById(id);
-        return ResponseEntity.ok(role);
-    }
-
-    @PostMapping
-    public ResponseEntity<Role> createRole(@RequestBody Role role) {
-        Role savedRole = roleService.saveRole(role);
-        return ResponseEntity.ok(savedRole);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteRole(@PathVariable Long id) {
-        roleService.deleteRole(id);
-        return ResponseEntity.noContent().build();
-    }
-}
