+++ src/main/java/ru/itmentor/spring/boot_security/demo/security/JWTAuthorizationFilter.java
@@ -1,4 +1,4 @@
-package ru.itmentor.spring.boot_security.demo.configs;
+package ru.itmentor.spring.boot_security.demo.security;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -18,29 +18,22 @@ public class JWTAuthorizationFilter extends OncePerRequestFilter {
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
             throws IOException, ServletException {
-
         String header = request.getHeader("Authorization");
-
         if (header == null || !header.startsWith("Bearer ")) {
             chain.doFilter(request, response);
             return;
         }
-
         String token = header.replace("Bearer ", "");
         try {
             String decoded = new String(Base64.getDecoder().decode(token));
             String username = decoded.split(":")[0];
-
             UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(
                     username, null, Arrays.asList(new SimpleGrantedAuthority("ROLE_ADMIN"))
             );
-
             SecurityContextHolder.getContext().setAuthentication(auth);
-
         } catch (Exception e) {
             SecurityContextHolder.clearContext();
         }
-
         chain.doFilter(request, response);
     }
 }
