+++ src/main/java/ru/itmentor/spring/boot_security/demo/security/LoginRequest.java
@@ -1,0 +1,22 @@
+package ru.itmentor.spring.boot_security.demo.security;
+
+public class LoginRequest {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
